agent/requirements.txt
beautifulsoup4>=4.12.0
google-genai>=0.5.0
google-auth>=2.35.0
gspread>=6.1.0
pypdf>=4.2.0
python-dotenv>=1.0.1
requests>=2.32.0
truststore>=0.10.0
